Move createInnerBlockList into `editor/utils`
